renamed SimpleLogic and AbstractLogic files according to the module names
