+++ src/veins/modules/world/traci/trafficLight/logics/TraCITrafficLightAbstractLogic.cc
@@ -18,7 +18,7 @@
 // Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 //
 
-#include "veins/modules/world/traci/trafficLight/logics/AbstractLogic.h"
+#include <veins/modules/world/traci/trafficLight/logics/TraCITrafficLightAbstractLogic.h>
 
 using Veins::TraCITrafficLightAbstractLogic;
 using namespace omnetpp;
