+++ src/veins/modules/world/traci/trafficLight/logics/TraCITrafficLightSimpleLogic.h
@@ -22,8 +22,8 @@
 #define TRACITRAFFICLIGHTSIMPLELOGIC_H_
 
 
+#include <veins/modules/world/traci/trafficLight/logics/TraCITrafficLightAbstractLogic.h>
 #include "veins/base/utils/FindModule.h"
-#include "veins/modules/world/traci/trafficLight/logics/AbstractLogic.h"
 #include "veins/modules/world/traci/trafficLight/TraCITrafficLightInterface.h"
